+++ src/sentry_transport.c
@@ -10,6 +10,7 @@ struct sentry_transport_s {
     int (*flush_func)(uint64_t timeout, void *state);
     void (*free_func)(void *state);
     size_t (*dump_func)(sentry_run_t *run, void *state);
+    void (*retry_func)(void *state);
     void (*cleanup_func)(const sentry_options_t *options, void *state);
     void *state;
     bool running;
@@ -147,6 +148,21 @@ void *
 sentry__transport_get_state(sentry_transport_t *transport)
 {
     return transport ? transport->state : NULL;
+}
+
+void
+sentry_transport_retry(sentry_transport_t *transport)
+{
+    if (transport && transport->retry_func) {
+        transport->retry_func(transport->state);
+    }
+}
+
+void
+sentry__transport_set_retry_func(
+    sentry_transport_t *transport, void (*retry_func)(void *state))
+{
+    transport->retry_func = retry_func;
 }
 
 void